refactor: general refactor of the migration app
Removed unused handlers and classes
Add saas/sm packages
Move handlers into correct packages
