--- migration/identity-migration/src/main/java/io/camunda/migration/identity/dto/UserGroups.java
@@ -1,12 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.migration.identity.dto;
-
-import java.util.List;
-
-public record UserGroups(String id, String username, List<Group> groups) {}
