--- migration/identity-migration/src/main/java/io/camunda/migration/identity/dto/UserResourceAuthorization.java
@@ -1,11 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.migration.identity.dto;
-
-public record UserResourceAuthorization(
-    String username, String resourceId, String resourceType, String permission) {}
