+++ migration/identity-migration/src/main/java/io/camunda/migration/identity/handler/saas/ClientMigrationHandler.java
@@ -5,17 +5,18 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.migration.identity;
+package io.camunda.migration.identity.handler.saas;
 
 import static io.camunda.migration.identity.config.saas.StaticEntities.getOperateClientPermissions;
 import static io.camunda.migration.identity.config.saas.StaticEntities.getTasklistClientPermissions;
 import static io.camunda.migration.identity.config.saas.StaticEntities.getZeebeClientPermissions;
 
 import io.camunda.migration.api.MigrationException;
-import io.camunda.migration.identity.console.ConsoleClient;
-import io.camunda.migration.identity.console.ConsoleClient.Client;
-import io.camunda.migration.identity.console.ConsoleClient.Members;
-import io.camunda.migration.identity.console.ConsoleClient.Permission;
+import io.camunda.migration.identity.client.ConsoleClient;
+import io.camunda.migration.identity.client.ConsoleClient.Client;
+import io.camunda.migration.identity.client.ConsoleClient.Members;
+import io.camunda.migration.identity.client.ConsoleClient.Permission;
+import io.camunda.migration.identity.handler.MigrationHandler;
 import io.camunda.security.auth.CamundaAuthentication;
 import io.camunda.service.AuthorizationServices;
 import io.camunda.service.AuthorizationServices.CreateAuthorizationRequest;
