+++ migration/identity-migration/src/main/java/io/camunda/migration/identity/handler/saas/StaticConsoleRoleMigrationHandler.java
@@ -5,13 +5,14 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.migration.identity;
+package io.camunda.migration.identity.handler.saas;
 
 import static io.camunda.migration.identity.config.saas.StaticEntities.ROLES;
 
 import io.camunda.migration.api.MigrationException;
-import io.camunda.migration.identity.console.ConsoleClient;
+import io.camunda.migration.identity.client.ConsoleClient;
 import io.camunda.migration.identity.dto.NoopDTO;
+import io.camunda.migration.identity.handler.MigrationHandler;
 import io.camunda.security.auth.CamundaAuthentication;
 import io.camunda.service.RoleServices;
 import io.camunda.service.RoleServices.RoleMemberRequest;
