--- migration/identity-migration/src/main/java/io/camunda/migration/identity/midentity/MigrationEntityType.java
@@ -1,17 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.migration.identity.midentity;
-
-public enum MigrationEntityType {
-  TENANT,
-  MAPPING_RULE,
-  TENANT_USER,
-  GROUP,
-  ROLE,
-  GROUP_USER
-}
